(PUP-5994) Fix module tar file extraction with long pathnames
The MiniTar extraction was failing intermittently for modules with long
filenames.  This was due to using entry.name intead of entry.full_name.
The full_name property is derived from the the filename prefix plus the
filename property.  By using the name property some files and directories
were silently dropped from the extraction.

This commit changes the validation to use full_name and adds a test to
extract an example tarball to enusre these files are indeed extracted.

Generally this commit will only affect the AIX, Solaris and Windows
operating systems due to the checking in 'lib/puppet/module_tool/tar.rb'
i.e. If the OS is not Windows and 'tar' is in the search path (which is
generally the case for unix based operating systems) then the Gnu class
is used, otherwise it will use the Minitar class if available.

Ustar Format:
https://en.wikipedia.org/wiki/Tar_(computing)#UStar_format
Field Offset 345 - Filename prefix

full_name function:
https://github.com/atoulme/minitar/blob/master/lib/archive/tar/minitar.rb#L547-L553